pyproject.toml
[project]
name = "monitoring-vercel"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "aiohttp==3.13.4",
    "websockets==12.0",
    "solders==0.23.0",
    "solana==0.36.2",
    "base58==2.1.1",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py39"
extend-exclude = [".claude"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "I",   # isort
    "B",   # bugbear
    "W",   # pycodestyle warnings
    "C",   # mccabe complexity
    "N",   # naming
    "D",   # pydocstyle
    "UP",  # pyupgrade
    "ANN", # annotations
    "RUF"  # ruff-specific rules
]
ignore = ["D203", "D213", "ANN401"]

[tool.ruff.lint.isort]
known-first-party = ["common", "metrics"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"test_*.py" = ["D", "ANN"]
"api/**/*.py" = ["N801"]

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true
disallow_untyped_defs = true